refactor(python): package root scripts under james_library
