+++ james_library/__init__.py
@@ -1,0 +1,1 @@
+"""Python package layout for james_library support modules."""
