+++ james_library/bootstrap/__init__.py
@@ -1,0 +1,1 @@
+"""Bootstrap and health-check modules for R.A.I.N. Lab."""
