+++ tests/test_external_integrations.py
@@ -1,10 +1,4 @@
-import sys
-from pathlib import Path
-
-
-sys.path.insert(0, str(Path(__file__).resolve().parent.parent))
-
-import external_integrations
+from james_library.services import external_integrations
 
 
 class _FakeResponse:
