+++ tests/test_tools_setup_code.py
@@ -1,11 +1,5 @@
 import ast
-import sys
-from pathlib import Path
-
-
-sys.path.insert(0, str(Path(__file__).resolve().parent.parent))
-
-import tools
+from james_library.utilities import tools
 
 
 def test_setup_code_is_valid_python():
